requirements.txt
# Core
fastapi
uvicorn[standard]
huey>=2.0
pydantic
PyYAML
openai

# Web & API
requests
httpx
cohere

# Data Processing
numpy
pandas
nltk
scikit-learn
nlpaug
protobuf==3.20.0

# PDF & Document Processing
pypdf
PyMuPDF
pytesseract
Pillow
beautifulsoup4
jinja2
python-docx

# Machine Learning & AI
# torch
transformers
datasets==2.15.0
# trl
# vllm
# unsloth
wandb

# Database & Caching
redis

# Utilities
tqdm
chardet
tabulate
matplotlib

certifi
chromadb
python-multipart
sentencepiece
pdf2image
runpod
rank_bm25
torch
peft
discord.py
